+++ src/tests/AutoSDK.SnapshotTests/Snapshots/elevenlabs/SystemTextJson/_#G.Models.ManualVerificationFileResponseModel.g.verified.cs
@@ -35,7 +35,7 @@ namespace G
         /// </summary>
         [global::System.Text.Json.Serialization.JsonPropertyName("size_bytes")]
         [global::System.Text.Json.Serialization.JsonRequired]
-        public required int SizeBytes { get; set; }
+        public required long SizeBytes { get; set; }
 
         /// <summary>
         /// The date of the file in Unix time.
@@ -75,7 +75,7 @@ namespace G
             string fileId,
             string fileName,
             string mimeType,
-            int sizeBytes,
+            long sizeBytes,
             int uploadDateUnix)
         {
             this.FileId = fileId ?? throw new global::System.ArgumentNullException(nameof(fileId));
