+++ src/tests/AutoSDK.SnapshotTests/Snapshots/elevenlabs/SystemTextJson/_#G.Models.RecordingResponseModel.g.verified.cs
@@ -28,7 +28,7 @@ namespace G
         /// </summary>
         [global::System.Text.Json.Serialization.JsonPropertyName("size_bytes")]
         [global::System.Text.Json.Serialization.JsonRequired]
-        public required int SizeBytes { get; set; }
+        public required long SizeBytes { get; set; }
 
         /// <summary>
         /// The date of the recording in Unix time.
@@ -74,7 +74,7 @@ namespace G
         public RecordingResponseModel(
             string recordingId,
             string mimeType,
-            int sizeBytes,
+            long sizeBytes,
             int uploadDateUnix,
             string transcription)
         {
