+++ src/tests/AutoSDK.SnapshotTests/Snapshots/github/SystemTextJson/_#G.Models.FullRepository.g.verified.cs
@@ -456,7 +456,7 @@ namespace G
         /// <example>108</example>
         [global::System.Text.Json.Serialization.JsonPropertyName("size")]
         [global::System.Text.Json.Serialization.JsonRequired]
-        public required int Size { get; set; }
+        public required long Size { get; set; }
 
         /// <summary>
         /// Example: master
@@ -1166,7 +1166,7 @@ namespace G
             int forksCount,
             int stargazersCount,
             int watchersCount,
-            int size,
+            long size,
             string defaultBranch,
             int openIssuesCount,
             bool hasIssues,
