+++ src/tests/AutoSDK.SnapshotTests/Snapshots/github/SystemTextJson/_#G.Models.MinimalRepository.g.verified.cs
@@ -437,7 +437,7 @@ namespace G
         /// The size of the repository, in kilobytes. Size is calculated hourly. When a repository is initially created, the size is 0.
         /// </summary>
         [global::System.Text.Json.Serialization.JsonPropertyName("size")]
-        public int? Size { get; set; }
+        public long? Size { get; set; }
 
         /// <summary>
         /// 
@@ -891,7 +891,7 @@ namespace G
             int? forksCount,
             int? stargazersCount,
             int? watchersCount,
-            int? size,
+            long? size,
             string? defaultBranch,
             int? openIssuesCount,
             bool? isTemplate,
