+++ src/tests/AutoSDK.SnapshotTests/Snapshots/openai/NewtonsoftJson/_#G.IUploadsClient.CreateUpload.g.verified.cs
@@ -64,7 +64,7 @@ namespace G
         global::System.Threading.Tasks.Task<global::G.Upload> CreateUploadAsync(
             string filename,
             global::G.CreateUploadRequestPurpose purpose,
-            int bytes,
+            long bytes,
             string mimeType,
             global::System.Threading.CancellationToken cancellationToken = default);
     }
